requirements.txt
python-dotenv==1.0.1
requests==2.32.3
feedparser==6.0.11
trafilatura==1.12.0
beautifulsoup4==4.12.3
Pillow==10.4.0
python-dateutil==2.9.0.post0
# Optional: improve extraction
# lxml>=5.1.0
